// src/connections/sunspec/models/inverter.ts
export type InverterModel = {
    /**
     * Model ID
     *
     * Model identifier
     */
    ID: 101 | 102 | 103;

    /**
     * Model Length
     *
     * Model length
     */
    L: number;

    /**
     * AC Current
     *
     * AC Current
     */
    A: number;

    /**
     * Phase A Current
     *
     * Phase A Current
     */
    AphA: number;

    /**
     * Phase B Current
     *
     * Phase B Current
     */
    AphB: number | null;

    /**
     * Phase C Current
     *
     * Phase C Current
     */
    AphC: number | null;

    /**
     * Scale Factor for Current
     */
    A_SF: number;

    /**
     * Phase Voltage AB
     *
     * Phase Voltage AB
     */
    PPVphAB: number | null;

    /**
     * Phase Voltage BC
     *
     * Phase Voltage BC
     */
    PPVphBC: number | null;

    /**
     * Phase Voltage CA
     *
     * Phase Voltage CA
     */
    PPVphCA: number | null;

    /**
     * Phase Voltage AN
     *
     * Phase Voltage AN
     */
    PhVphA: number;

    /**
     * Phase Voltage BN
     *
     * Phase Voltage BN
     */
    PhVphB: number | null;

    /**
     * Phase Voltage CN
     *
     * Phase Voltage CN
     */
    PhVphC: number | null;

    /**
     * Scale Factor for Voltage
     */
    V_SF: number;

    /**
     * AC Power
     *
     * AC Power
     */
    W: number;

    /**
     * Scale Factor for Power
     */
    W_SF: number;

    /**
     * Line Frequency
     *
     * Line Frequency
     */
    Hz: number;

    /**
     * Scale Factor for Frequency
     */
    Hz_SF: number;

    /**
     * AC Apparent Power
     *
     * AC Apparent Power
     */
    VA: number | null;

    /**
     * Scale Factor for Apparent Power
     */
    VA_SF: number | null;

    /**
     * AC Reactive Power
     *
     * AC Reactive Power
     */
    VAr: number | null;

    /**
     * Scale Factor for Reactive Power
     */
    VAr_SF: number | null;

    /**
     * AC Power Factor
     *
     * AC Power Factor
     */
    PF: number | null;

    /**
     * Scale Factor for Power Factor
     */
    PF_SF: number | null;

    /**
     * AC Energy
     *
     * AC Energy
     */
    WH: number;

    /**
     * Scale Factor for Energy
     */
    WH_SF: number;

    /**
     * DC Amps
     *
     * DC Current
     */
    DCA: number | null;

    /**
     * Scale Factor for DC Current
     */
    DCA_SF: number | null;

    /**
     * DC Voltage
     *
     * DC Voltage
     */
    DCV: number | null;

    /**
     * Scale Factor for DC Voltage
     */
    DCV_SF: number | null;

    /**
     * DC Watts
     *
     * DC Power
     */
    DCW: number | null;

    /**
     * Scale Factor for DC Power
     */
    DCW_SF: number | null;

    /**
     * Cabinet Temperature
     *
     * Cabinet Temperature
     */
    TmpCab: number | null;

    /**
     * Heat Sink Temperature
     *
     * Heat Sink Temperature
     */
    TmpSnk: number | null;

    /**
     * Transformer Temperature
     *
     * Transformer Temperature
     */
    TmpTrns: number | null;

    /**
     * Other Temperature
     *
     * Other Temperature
     */
    TmpOt: number | null;

    /**
     * Scale Factor for Temperature
     */
    Tmp_SF: number | null;

    /**
     * Operating State
     *
     * Enumerated value. Operating state
     */
    St: InverterState;

    /**
     * Vendor Operating State
     *
     * Vendor specific operating state code
     */
    StVnd: number | null;

    /**
     * Event1
     *
     * Bitmask value. Event fields
     */
    Evt1: InverterEvent1;

    /**
     * Event Bitfield 2
     *
     * Reserved for future use
     */
    Evt2: number;

    /**
     * Vendor Event Bitfield 1
     *
     * Vendor defined events
     */
    EvtVnd1: number | null;

    /**
     * Vendor Event Bitfield 2
     *
     * Vendor defined events
     */
    EvtVnd2: number | null;

    /**
     * Vendor Event Bitfield 3
     *
     * Vendor defined events
     */
    EvtVnd3: number | null;

    /**
     * Vendor Event Bitfield 4
     *
     * Vendor defined events
     */
    EvtVnd4: number | null;
};
// ...
export enum InverterState {
    // Off
    OFF = 1,
    // Sleeping (auto-shutdown)
    SLEEPING = 2,
    // Starting up
    STARTING = 3,
    // Tracking power point
    MPPT = 4,
    // Forced power reduction
    THROTTLED = 5,
    // Shutting down
    SHUTTING_DOWN = 6,
    // One or more faults exist
    FAULT = 7,
    // Standby (service on unit)* might be in Events
    STANDBY = 8,
}

/**
 * InverterEvent1 Bitfield Enumeration
 *
 * Bitmask values representing various event fields
 */
export enum InverterEvent1 {
    GROUND_FAULT = 1 << 0,
    DC_OVER_VOLT = 1 << 1,
    AC_DISCONNECT = 1 << 2,
    DC_DISCONNECT = 1 << 3,
    GRID_DISCONNECT = 1 << 4,
    CABINET_OPEN = 1 << 5,
    MANUAL_SHUTDOWN = 1 << 6,
    OVER_TEMP = 1 << 7,
    OVER_FREQUENCY = 1 << 8,
    UNDER_FREQUENCY = 1 << 9,
    AC_OVER_VOLT = 1 << 10,
    AC_UNDER_VOLT = 1 << 11,
    BLOWN_STRING_FUSE = 1 << 12,
    UNDER_TEMP = 1 << 13,
    MEMORY_LOSS = 1 << 14,
    HW_TEST_FAILURE = 1 << 15,
}